[refactor] user와 diag 분리
