+++ diag/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from diag.views import *
+
+urlpatterns = [
+    path('/record', RecordView.as_view()),
+]
