+++ users/models.py
@@ -25,14 +25,3 @@ TEST_USER_DATA = {
     "password": "test",
     "username": "test"
 }
-
-class DiagRecord(BaseModel):
-    totalQuestionSize = models.IntegerField()
-    yesCount = models.IntegerField()
-
-    user = models.ForeignKey('users.User', related_name='diagrecord', on_delete=models.CASCADE)
-
-TEST_DIAG_DATA = {
-    "totalQuestionSize": 12,
-    "yesCount": 3,
-}
