+++ users/urls.py
@@ -1,8 +1,6 @@
 from django.urls import path
 
 from users.views.user_check_views import *
-
-from users.views.user_diag_views import *
 from users.views.user_get_update_delete_views import *
 from users.views.user_login_views import *
 from users.views.user_join_views import *
@@ -14,6 +12,4 @@ urlpatterns = [
     path('/<int:userId>/auto/login', AutoLoginView.as_view()),
     path('/<int:userId>', GetUserAndDeleteView.as_view()),
     path('/validate/accountId', CheckAccountIdView.as_view()),
-    path('/recordsave', RecordSaveView.as_view()),
-    path('/prevrecord', GetDiagRecordView.as_view()),
 ]
